feat: add CSS color resolution for PDF exports, update security response analyzer logic, and include userApiCallStatuses in evaluation payload.
