+++ backend/src/inngest/services.ts
@@ -890,7 +890,7 @@ export async function processAutomatedApiTest(
   await step.run("initialize-user-api-responses", async () => {
     await pool.query(
       `UPDATE evaluation_status
-       SET payload = COALESCE(payload, '{}'::jsonb) || '{"userApiResponses": []}'::jsonb
+       SET payload = COALESCE(payload, '{}'::jsonb) || '{"userApiResponses": [], "userApiCallStatuses": {}}'::jsonb
        WHERE id = $1`,
       [job.id]
     );
