+++ frontend/src/app/assess/[projectId]/fairness-bias/dataset-testing/report/hooks/usePdfExport.ts
@@ -8,59 +8,12 @@ import {
     removeAnimations, fixSensitiveAnalysis
 } from "./pdfStyles";
 import { THRESHOLDS } from "../../constants";
-
-const oklchCache = new Map<string, string>();
-const convertOklchToRgb = (colorStr: string): string => {
-    if (oklchCache.has(colorStr)) {
-        return oklchCache.get(colorStr)!;
-    }
-    try {
-        if (typeof document === "undefined") return "rgb(0, 0, 0)";
-        const canvas = document.createElement("canvas");
-        canvas.width = 1;
-        canvas.height = 1;
-        const ctx = canvas.getContext("2d");
-        if (!ctx) return "rgb(0, 0, 0)";
-        ctx.fillStyle = colorStr;
-        const resolved = ctx.fillStyle;
-        oklchCache.set(colorStr, resolved);
-        return resolved;
-    } catch (e) {
-        return "rgb(0, 0, 0)";
-    }
-};
-
-const resolveColorFunctions = (value: any): any => {
-    if (typeof value !== "string") return value;
-    if (!value.includes("oklch") && !value.includes("oklab") && !value.includes("lab") && !value.includes("lch")) {
-        return value;
-    }
-    return value.replace(/(oklch|oklab|lab|lch)\([^)]+\)/g, (match) => {
-        return convertOklchToRgb(match);
-    });
-};
-
-const ruleCache = new WeakMap<any, any>();
-const getProxyRule = (rule: any): any => {
-    if (!rule) return rule;
-    if (ruleCache.has(rule)) return ruleCache.get(rule);
-
-    const proxy = new Proxy(rule, {
-        get(target, prop, receiver) {
-            if (prop === "cssText") {
-                try {
-                    const originalText = target.cssText;
-                    return resolveColorFunctions(originalText);
-                } catch (e) {
-                    return Reflect.get(target, prop, receiver);
-                }
-            }
-            return Reflect.get(target, prop, receiver);
-        }
-    });
-    ruleCache.set(rule, proxy);
-    return proxy;
-};
+import { 
+    getStyleProxy, 
+    getProxyRule, 
+    acquireExportLock, 
+    releaseExportLock 
+} from "@/lib/pdfExport/pdfColorResolver";
 
 interface UsePdfExportProps {
     reportRef: RefObject<HTMLDivElement>;
@@ -87,6 +40,12 @@ export const usePdfExport = ({ reportRef, payload }: UsePdfExportProps) => {
     const exportPdf = useCallback(async () => {
         if (!reportRef.current || !payload) return;
         if (isExportingRef.current) return;
+
+        const acquired = await acquireExportLock();
+        if (!acquired) {
+            console.warn("Another PDF export is currently in progress. Aborting.");
+            return;
+        }
 
         let originalDescriptor: any = null;
         let originalGetComputedStyle: any = null;
@@ -142,21 +101,11 @@ export const usePdfExport = ({ reportRef, payload }: UsePdfExportProps) => {
                 originalGetComputedStyle = window.getComputedStyle;
                 window.getComputedStyle = function(element, pseudoElt) {
                     const style = originalGetComputedStyle.call(this, element, pseudoElt);
-                    return new Proxy(style, {
-                        get(target, prop, receiver) {
-                            if (prop === "getPropertyValue") {
-                                return function(propertyName: string) {
-                                    const val = target.getPropertyValue(propertyName);
-                                    return resolveColorFunctions(val);
-                                };
-                            }
-                            const val = Reflect.get(target, prop, receiver);
-                            if (typeof val === "string") {
-                                return resolveColorFunctions(val);
-                            }
-                            return val;
-                        }
-                    });
+                    const isInsidePdfContainer = element.closest("#pdf-export-container") || element.id === "pdf-export-container";
+                    if (isInsidePdfContainer) {
+                        return getStyleProxy(style);
+                    }
+                    return style;
                 };
             }
 
@@ -1003,6 +952,7 @@ export const usePdfExport = ({ reportRef, payload }: UsePdfExportProps) => {
 
             setIsExporting(false);
             isExportingRef.current = false;
+            releaseExportLock();
         }
     }, [reportRef, payload]);
