+++ frontend/src/app/assess/[projectId]/vulnerability-assessment/api-history/[reportId]/page.tsx
@@ -125,9 +125,8 @@ export default function VulnerabilityReportDetailPage() {
         };
     }, [reportId]);
 
-    const getScoreColor = (score: number) => {
-        // Normalize score to 0-100 if it's on a 0-1 scale
-        const normalizedScore = score <= 1 ? score * 100 : score;
+    const getScoreColor = (score: number, isFraction: boolean = false) => {
+        const normalizedScore = isFraction ? score * 100 : score;
         if (normalizedScore >= 80) return "text-emerald-500";
         if (normalizedScore >= 60) return "text-amber-500";
         return "text-red-500";
@@ -262,7 +261,7 @@ export default function VulnerabilityReportDetailPage() {
                                     {isSecurityReport ? "Avg Overall Score" : "Avg Fairness Score"}
                                 </div>
                                 <div className={`text-3xl font-black ${report.average_scores?.averageOverallScore != null
-                                    ? getScoreColor(report.average_scores.averageOverallScore)
+                                    ? getScoreColor(report.average_scores.averageOverallScore, true)
                                     : "text-muted-foreground"
                                     }`}>
                                     {report.average_scores?.averageOverallScore != null
@@ -275,7 +274,7 @@ export default function VulnerabilityReportDetailPage() {
                                     {isSecurityReport ? "Avg Vulnerability Score" : "Avg Bias Score"}
                                 </div>
                                 <div className={`text-3xl font-black ${report.average_scores?.averageBiasScore != null
-                                    ? getScoreColor(1 - (report.average_scores.averageBiasScore ?? 0))
+                                    ? getScoreColor(1 - (report.average_scores.averageBiasScore ?? 0), true)
                                     : "text-muted-foreground"
                                     }`}>
                                     {report.average_scores?.averageBiasScore != null
@@ -513,7 +512,7 @@ export default function VulnerabilityReportDetailPage() {
                                                                 <div className="flex justify-between items-center text-sm">
                                                                     <span className="text-muted-foreground">Overall Score:</span>
                                                                     <span className={`font-black ${Number.isFinite((item as any).evaluation?.overallScore)
-                                                                        ? getScoreColor((item as any).evaluation?.overallScore)
+                                                                        ? getScoreColor((item as any).evaluation?.overallScore, true)
                                                                         : "text-muted-foreground"
                                                                         }`}>
                                                                         {Number.isFinite((item as any).evaluation?.overallScore)
