+++ frontend/src/hooks/usePdfReport.ts
@@ -9,127 +9,12 @@ import {
 import { pdf } from "@react-pdf/renderer";
 import { AimaPdfDocument } from "../lib/pdfExport/AimaPdfDocument";
 
-const oklchCache = new Map<string, string>();
-const convertOklchToRgb = (colorStr: string): string => {
-    if (oklchCache.has(colorStr)) {
-        return oklchCache.get(colorStr)!;
-    }
-    try {
-        if (typeof document === "undefined") return "rgb(0, 0, 0)";
-        const canvas = document.createElement("canvas");
-        canvas.width = 1;
-        canvas.height = 1;
-        const ctx = canvas.getContext("2d");
-        if (!ctx) return "rgb(0, 0, 0)";
-        ctx.fillStyle = colorStr;
-        const resolved = ctx.fillStyle;
-        oklchCache.set(colorStr, resolved);
-        return resolved;
-    } catch (e) {
-        return "rgb(0, 0, 0)";
-    }
-};
-
-const resolveColorFunctions = (value: any): any => {
-    if (typeof value !== "string") return value;
-    if (!value.includes("oklch") && !value.includes("oklab") && !value.includes("lab") && !value.includes("lch")) {
-        return value;
-    }
-    return value.replace(/(oklch|oklab|lab|lch)\([^)]+\)/g, (match) => {
-        return convertOklchToRgb(match);
-    });
-};
-
-const styleProxyCache = new WeakMap<any, any>();
-const getStyleProxy = (style: any): any => {
-    if (!style) return style;
-    if (styleProxyCache.has(style)) return styleProxyCache.get(style);
-
-    const proxy = new Proxy(style, {
-        get(target, prop) {
-            if (prop === "getPropertyValue") {
-                return function(propertyName: string) {
-                    const val = target.getPropertyValue(propertyName);
-                    return resolveColorFunctions(val);
-                };
-            }
-            let val = Reflect.get(target, prop);
-            if (typeof val === "function") {
-                return val.bind(target);
-            }
-            if (typeof val === "string") {
-                return resolveColorFunctions(val);
-            }
-            return val;
-        }
-    });
-    styleProxyCache.set(style, proxy);
-    return proxy;
-};
-
-const ruleCache = new WeakMap<any, any>();
-const getProxyRule = (rule: any): any => {
-    if (!rule) return rule;
-    if (ruleCache.has(rule)) return ruleCache.get(rule);
-
-    const proxy = new Proxy(rule, {
-        get(target, prop) {
-            if (prop === "cssText") {
-                try {
-                    const originalText = target.cssText;
-                    return resolveColorFunctions(originalText);
-                } catch (e) {
-                    let val = Reflect.get(target, prop);
-                    if (typeof val === "function") return val.bind(target);
-                    return val;
-                }
-            }
-            if (prop === "style") {
-                return getStyleProxy(target.style);
-            }
-            if (prop === "cssRules") {
-                try {
-                    const rules = target.cssRules;
-                    if (!rules) return rules;
-                    const filteredRules: any[] = [];
-                    for (let i = 0; i < rules.length; i++) {
-                        filteredRules.push(getProxyRule(rules[i]));
-                    }
-                    const ruleList: any = {
-                        length: filteredRules.length,
-                        item(index: number) { return filteredRules[index]; },
-                        [Symbol.iterator]() {
-                            let i = 0;
-                            return {
-                                next() {
-                                    return i < filteredRules.length
-                                        ? { value: filteredRules[i++], done: false }
-                                        : { done: true };
-                                }
-                            };
-                        }
-                    };
-                    filteredRules.forEach((rule, idx) => {
-                        ruleList[idx] = rule;
-                    });
-                    if (typeof CSSRuleList !== "undefined") {
-                        Object.setPrototypeOf(ruleList, CSSRuleList.prototype);
-                    }
-                    return ruleList;
-                } catch (e) {
-                    // fallback
-                }
-            }
-            let val = Reflect.get(target, prop);
-            if (typeof val === "function") {
-                return val.bind(target);
-            }
-            return val;
-        }
-    });
-    ruleCache.set(rule, proxy);
-    return proxy;
-};
+import { 
+    getStyleProxy, 
+    getProxyRule, 
+    acquireExportLock, 
+    releaseExportLock 
+} from "../lib/pdfExport/pdfColorResolver";
 
 interface PdfExportOptions {
     reportRef: RefObject<HTMLDivElement>;
@@ -302,6 +187,12 @@ export const usePdfReport = ({
         if (!reportRef.current) return;
         if (isExportingRef.current) return;
 
+        const acquired = await acquireExportLock();
+        if (!acquired) {
+            console.warn("Another PDF export is currently in progress. Aborting.");
+            return;
+        }
+
         let originalDescriptor: any = null;
         let originalGetComputedStyle: any = null;
 
@@ -356,7 +247,11 @@ export const usePdfReport = ({
                 originalGetComputedStyle = window.getComputedStyle;
                 window.getComputedStyle = function(element, pseudoElt) {
                     const style = originalGetComputedStyle.call(this, element, pseudoElt);
-                    return getStyleProxy(style);
+                    const isInsidePdfContainer = element.closest("#pdf-export-container") || element.id === "pdf-export-container";
+                    if (isInsidePdfContainer) {
+                        return getStyleProxy(style);
+                    }
+                    return style;
                 };
             }
             
@@ -383,7 +278,7 @@ export const usePdfReport = ({
             const pageHeight = pdf.internal.pageSize.getHeight(); // 297mm
             const { margin, headerHeight, footerHeight } = PDF_CONFIG;
             const usableWidth = pageWidth - 2 * margin; // 190mm
-            const contentTop = 30; // Start below the 28mm header
+            const contentTop = headerHeight + 2; // Start below the header with 2mm buffer
             const contentBottom = pageHeight - footerHeight - 5;
             const usableHeight = contentBottom - contentTop;
 
@@ -391,11 +286,11 @@ export const usePdfReport = ({
             const addPageHeader = (pdfDoc: any, pageNum: number, totalPages: number) => {
                 // Clear the header area with white rect first to avoid content bleed
                 pdfDoc.setFillColor(255, 255, 255);
-                pdfDoc.rect(0, 0, pageWidth, 28, "F");
+                pdfDoc.rect(0, 0, pageWidth, headerHeight, "F");
 
                 // FULL BLEED Header background - our primary blue (#0070ea)
                 pdfDoc.setFillColor(0, 112, 234); 
-                pdfDoc.rect(0, 0, pageWidth, 28, "F");
+                pdfDoc.rect(0, 0, pageWidth, headerHeight, "F");
 
                 // Report title - centered white text
                 pdfDoc.setFont("helvetica", "bold");
@@ -656,6 +551,7 @@ export const usePdfReport = ({
             }
             setIsExporting(false);
             isExportingRef.current = false;
+            releaseExportLock();
         }
     }, [reportRef, fileName, reportTitle, projectName, generatedAt, sectionSelector]);
